Email: Make the language of folder name show rightly in the widget
Email has saved the string of the folder name by add the attribute in the
xml file, when we change the language, the folder name can't change as the
system language.

We have delete the attribute in the xml so that the language will change
as the system language.

CRs-fixed: 784554

Change-Id: I47951bce42796772b81349ddbda67ba187e0138b