UnifiedEmail: Modify to make email not crash when view an image.
When we view an image, it will use intent to save message content,
and it will cause TransactionTooLargeException when the content is
too large.

Because of the content is not used when we save it in the content,
so now we don't save the content in the intent when we view an image.

Change-Id: I3b257bb5cb098932520ec6855ed7c69b33f42148
CRs-fixed: 948500