UnifiedEmail: fix repeated attachments
When forwarding an email, its attachments will be automatically added to
the draft. If they are then removed or changed, subsequent mode switches
will repeatedly add the original attachments to the draft. Don't
reinitialize attachments from the original mail in this case, as they
won't be removed in resetMessageForModeChange().

Change-Id: Ibeedd508dfe97036a3eba3d9d89eb763b6179ae7
CRs-Fixed: 2062458