UnifiedEmail: fix IllegalStateException
An IllgalStateException occured during monkey test due to committing
Fragment transaction after onSaveInstanceState(). The offending code
is indeed executed on UI thread, so either the click event is fired
after onPause(), or onSaveIntanceState() is called by framework before
onPause(). Use commitAllowingStateLoss() to avoid crash. This won't
affect user because it doesn't make sense anyway to show DialogFragment
when the Activity is not in foreground.

Change-Id: Ifd7de008b473405a379811dde78d003740436509
CRs-Fixed: 2068335