Email: Fix FC when adding attachment twice
FC occurs when user added attachment and try to add a new
one. Root cause is that attachment preview image will be
saved when go to DocumentsUI at the second time. The image
data is large, hence it causes TransactionLargeException.
So don't save the preview image since it's unnecessary.

CRs-Fixed: 2236176

Change-Id: I05220dafa69760dde9f00daace0897ce07cce668